provide mock results generator (&data) for testing
Due to current privacy concerns, no real results are committed or stored on the heroku server. This commit adds a python script to generate mock data, an example of which is committed ("data/results.json").
If "real" data is available at "./dataPrivate/example-data-export.json" (which is gitignored) then it is preferentially sourced by the server.
Reducer & action names changed accordingly.