Implement basic authentication
Here we implement a basic login functionality where the client sends a username & password (always over HTTPS) and the server responds with a JWT which the client stores in `localStorage`.

The available username, password & JWT secret are currently sourced from env variables (by the server). This is only acceptable for a prototype (see README).

Not yet implemented:
* JWT expiry
* login rate limiting & max retries
* JWT used for API authentication (the point!)
* ability to revoke JWTs & client check JWT hasn't been revoked
* decode JWT in client to get username
* a development flag to ignore authentication (!)
* store hashed passwords (server) -- see note in server/auth.js
* properly test JWT signature (currently using default HS256)