geo scale now always shows 0-100%
Previously each chart (table & geo) gets it's own slice of the (transformed) data and sets its scales accordingly. When faceting, however, each scale would be different but use the same colour scale, causing confusing results. Fixing this means modifying the selectors to make the scales (e.g. maximum x-value) for all charts available to a higher component, which would then pass that information to each chart. This commit sidesteps this change by fixing all geo scales at 0-100%, and choosing a colour scale with more power to show the greater ranges which result. The tables are not modified here, but their x-axis are different when faceting which is equally confusing.