allow transitions between percentage & counts on x-axis of table
The aesthetic change is the ability to toggle the x-axis value and have a smooth transition between the two. This necessitated refactoring of the `useEffect()` hook - specifically the design of passing in a `ref` object (from `useRef()`) allowing us to store values used during the last render so that we can work out what to change.