Add pathogen to mock data & enable pathogen sidebar selector
Implementing this improved the "results" selectors (but there is still a lot of room for improvement there). Pathogen dropdown now filters appropriately, and pathogen is available as a primary & groupBy variable.

As the changing of pathogen is simply a filtering of the data, the table can transition between these data the same way as the count-perc toggle.