fix deps problem with useEffect hook
Previously we ignored some of the deps to the <Geo> component's `useEffect` hook -- never a good idea. This was done to avoid rerendering the map every time `mouseenter` fired, as that would update the both the hoverState and the handlers. This solution wraps the handlers in `useCallback` hooks ensuring they don't change.