Check input data for errors & transform as needed
Here we implement (somewhat incomplete) data checking and transforming as we parse the data, which is then passed to the reducer to store in state. This provides a singular point where data is transformed from the format it arrives in to the format the app uses it in. This will be very helpful if we ever change the input data shape.