Visualise modelling results
We can now visualise modeling results by changing the "data source" dropdown. The client sends an API request to the server, which currently uses 2 hardcoded CSVs taken from the simulated-data repo. This is temporary and should be replaced with the modeling server shortly.

Note that there is a race condition bug in the middleware used to fetch the modeling results. See the comment in the code.