Prototype encoding absolute count to circle size
I believe the chloropleths were not presenting the core thing they needed to present, that is the count of encounters in a region. Incidence is the primary data variable we want to display. Sex, age, vaccination, etc... are secondary colorings.

This keeps the same color encoding as previously but rather than coloring the full geoJSON path, we instead color a circle. This circle is sized according to absolute total count from this region.