Add Incidence as a primary variable
Previously, what we're calling "Incidence" (but is really
just count of participants) was missing from the viz
prototype. Add a temporary solution to make 
"Incidence" work with the existing viz.

I describe this commit as "temporary" because all
these if/else and switch/case statements are red flags
to me, signaling that the architecture of this app
should be rewritten, perhaps in an OO-design.