WIP: connect viz to modeled results
This commit partially connects the viz prototype with the IDM API
through a proxy GET endpoint that we host as part of our API.

The IDM API cannot be fully integrated with this viz until further
design questions are considered. There is a lot of data standardization
that needs to be done between IDM and ID3C to avoid lots of if/else or
branching logic as part of this app.

Currently, the prototype is able to get modeled results for `all`
pathogens as well as Flu A H1 & H3. We are getting results for Seattle
neighborhoods only. We are not able to query or filter by age, sex, or
vaccination status. The time (Week) filter does not currently work in
the app, but it should be able to be integrated with React with the
current data. 'Mean' is not an available result for every pathogen
model.

NOTE:
The bar chart and the map are not rendering the information correctly.
This is still a work in progress that is being put aside until we
collaborate further on the future of the IDM modeling results API.

Please see [this GitHub
issue](https://github.com/seattleflu/incidence-mapper/issues/87) for
more context.