api: create GET endpoint for model results
For testing purposes (and also to get around the lack 
of CORS policy on IDM's side), create a GET endpoint
on our server to query the modeled incidence 
results by pathogen. 

This query is likely to change, and eventually, we may
even remove this GET endpoint and query the IDM
API directly via React.